requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
python-decouple==3.8